# backend/routes/spec/api-docs.yaml
openapi: "3.0.0"
info:
  version: 0.0.1
  title: Digital Realty Website Backend API
  description: |
    Provides the backend/bridge for the digital realty website
  contact:
    name: Propane Studios
servers:
  - url: /api/v1
    description: Version 1 of this server
security:
  - basicAuth: []
paths:
...
  /Icon:
    get:
      description: Get a list of all icons.
      summary: List all Icons
      operationId: ListIcons
      tags:
        - Icon
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Limit"
        - name: page
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Page"
      responses:
           '200':
            description: A paged array of Icon
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    icon:
                      type: array
                      items:
                        $ref: "#/components/schemas/Icon"
                    total:
                      $ref: "#/components/schemas/Total"
                    numPages:
                      $ref: "#/components/schemas/NumPages"
           '401':
              description: Unauthorized
              content:
                application/json:
                  schema:
                    $ref: "#/components/schemas/NotAuthorized"
           '404':
              description: Not Found
              content:
                application/json:
                  schema:
                    $ref: "#/components/schemas/NotFound"
           '500':
              description: unexpected error
              content:
                application/json:
                  schema:
                    $ref: "#/components/schemas/Error"
    post:
      description: Create Icon
      operationId: createIcon
      summary: Create Icon
      tags:
      - Icon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostIcon"
      responses:
        '201':
          description: Icon created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessIcon"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidRequest"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAuthorized"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
        '500':
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /Icon/{id}:
    get:
      description: Get a list of all Icons where the role of solutions architect or admin
      summary: Get an Icon
      tags:
        - Icon
      parameters:
        - name: id
          in: path
          dscription: The Icon Id
          required: true
          schema:
            type: string
          example: 1a7bca18-75f0-4c29-b82b-9d92bb778007
      responses:
        '200':
          description: A list of Icon that the logged in user has access to
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Icon"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAuthorized"
        '404':
          description: not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      description: updates the Icon. The logged in user must be in role solutions architect
      operationId: update Icon
      summary: used to update data of an Icon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/IdFields'
      tags:
        - Icon
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              $ref: "#/components/schemas/PutIcon"
      responses:
        '200':
          description: Successfully updated Icon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Icon"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidRequest"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAuthorized"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
        '500':
          description: error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /Icon/{id}/activate:
    put:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/IdFields'
      tags:
        - Icon
      summary: used to activate icon
      responses:
        '200':
          description: Successfully unarchived (activated)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Icon"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidRequest"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAuthorized"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
        '500':
          description: error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/IdFields'
      tags:
        - Icon
      summary: Icon archived
      responses:
        '200':
          description: Successfully archived icon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Icon"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidRequest"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAuthorized"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
        '500':
          description: error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    Total:
      type: number
      description: total number of records
    NumPages:
      type: number
      description: total number of pages given the limit
    Limit:
      type: number
      description: Total number of records per page (max 100)
    Page:
      description: Which page to start on (relative to limit)
      type: integer
      format: int32
...
    InvalidRequest:
      required:
        - error
      properties:
        error:
          type: string
          description: Contains a message detailing why the request was invalid
    Error:
      required:
        - error
      properties:
        error:
          type: string
          description: contains a message about the failure
...
    CreatedBy:
      type: string
      description: id for creating user
      example: b0d0ab14-4069-48df-9258-86647a0d7d95
    IdFields:
      type: string
      description: Object id (UUID)
      example: b0d0ab14-4069-48df-9258-86647a0d7d95
...
    ArchivedFields:
      type: boolean
      description: Whether or not this is archived
      example: false
...
    CreatedAtFields:
      type: string
      description: date the record was created
      example: '2022-09-16T13:00:00.000Z'
    UpdatedAtFields:
      type: string
      description: date the record was updated
      example: '2022-09-16T14:00:00.000Z'
...
    Icon:
      required:
        - Id
        - iconName
        - createdBy
        - archived
        - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/IdFields'
        iconName:
          $ref: '#/components/schemas/IconNameField'
        createdBy:
          $ref: '#/components/schemas/CreatedAtFields'
        archived:
          $ref: '#/components/schemas/ArchivedFields'
        updatedAt:
          $ref: '#/components/schemas/UpdatedAtFields'
    PostIcon:
      required:
        - iconName
      properties:
        iconName:
          $ref: '#/components/schemas/IconNameField'
    PutIcon:
      properties:
        id:
          $ref: '#/components/schemas/IdFields'
        industryVertical:
          $ref: '#/components/schemas/IconIndustryVerticalField'
        useCase:
          $ref: '#/components/schemas/IconUseCaseField'
        notes:
          $ref: '#/components/schemas/IconNotesField'
        docName:
          $ref: '#/components/schemas/IconNameField'
        types:
          $ref: '#/components/schemas/IconOriginalFileNameField'
        createdBy:
          $ref: '#/components/schemas/CreatedBy'
        archived:
          $ref: '#/components/schemas/ArchivedFields'
        updatedAt:
          $ref: '#/components/schemas/UpdatedAtFields'
        createdAt:
          $ref: '#/components/schemas/CreatedAtFields'
        storageLocation:
          $ref: '#/components/schemas/IconStorageLocationField'
    IconIndustryVerticalField:
      type: string
      description: Industry Vertical name
      example: Aritificial Intelligence
    IconUseCaseField:
      type: string
      description: Industry Vertical name
      example: Aritificial Intelligence
    IconNameField:
      type: string
      description: Icon Name 
      example: Add name
    IconNotesField:
      type: string
      description: Notes name
      example: Add Notes
    IconOriginalFileNameField:
      type: string
      description:
      example: example message
    IconStorageLocationField:
      type: string
      description: Icon storage location name
      example: example emessage
    SuccessIcon:
      required:
        - iconName
      properties:
        docName:
          type: object
          description: Contains the success message
          required:
            - id
            - industryVertical
            - useCase
            - iconName
            - originalFilename
            - createdBy
            - archived
            - updatedAt
            - createdAt
            - storageLocation
            - notes
          properties:
            id:
              $ref: '#/components/schemas/IdFields'
            industryVertical:
              $ref: '#/components/schemas/IconIndustryVerticalField'
            useCase:
              $ref: '#/components/schemas/IconUseCaseField'
            iconName:
              $ref: '#/components/schemas/IconNameField'
            originalFilename:
              $ref: '#/components/schemas/IconOriginalFileNameField'
            createdBy:
              $ref: '#/components/schemas/CreatedBy'
            archived:
              $ref: '#/components/schemas/ArchivedFields'
            updatedAt:
              $ref: '#/components/schemas/UpdatedAtFields'
            createdAt:
              $ref: '#/components/schemas/CreatedAtFields'
            storageLocation:
              $ref: '#/components/schemas/IconStorageLocationField'
            notes:
              $ref: '#/components/schemas/IconNotesField'
    NotFound:
      required:
        - error
      properties:
        error:
          type: string
          example: "Not Found"
    NotAuthorized:
      required:
        - error
      properties:
        error:
          type: string
          example: "Not Authorized"